This code will get every single EPIC ID in Campaign 2 and put it into a single, gigantic dataframe called information1. If you want the EPIC ID's organized by which program they appear in, they are accessible via a series called information. There are many redundant entries that I have yet to clean out. The number of EPIC ID's in information1 is about 29 thousand. The number of unique EPIC ID's in all of Campaign 2 is 16622. The reason for the excess values is that some EPIC ID's are used in multiple programs.
There is a command called drop_duplicates that should clean out the duplicated EPIC ID's, but it reduces the number of EPIC ID's to about 13 thousand. I'm not sure why the drop_duplicate function takes away some of the unique values. If anyone has ideas on how to efficiently delete the correct number of redundant entries, it would be very helpful. I would also like to keep track of which programs the EPIC ID's are used in (we may need that data at some point), but that is not necessarily as important as just cleaning out redundant entries from the dataframe.

I would recommend that you read up on the documentation of the pandas library if you are unfamiliar with series and dataframes.